Small cleanup, integrate changes into kg_main for end to end
